UPDATE: Atualizando os testes de relatórios
